requirements.txt
aiohttp
jinja2
python-dotenv
firebase-admin
pymupdf
pillow
